refactor(chat): フォルダ構成の整理
