+++ apps/web/src/features/chat/components/parts/message-part.tsx
@@ -2,11 +2,11 @@ import type { UIMessage } from '@tanstack/ai'
 import type { AnyClientTool, MessagePart } from '@tanstack/ai-client'
 import { FileTextIcon } from 'lucide-react'
 
-import { isChatAttachmentTextPart } from '../hooks/use-chat-attachment'
-import type { ChatAttachmentMetadata } from '../hooks/use-chat-attachment'
-import { formatBytes } from '../utils/format-bytes'
-import { ExpandableSection } from './expandable-section'
-import { ParameterDisplay } from './parameter-display'
+import { isChatAttachmentTextPart } from '../../hooks/use-chat-attachment'
+import type { ChatAttachmentMetadata } from '../../hooks/use-chat-attachment'
+import { formatBytes } from '../../utils/format-bytes'
+import { ExpandableSection } from '../utils/expandable-section'
+import { ParameterDisplay } from '../utils/parameter-display'
 import { TextContent } from './text-content'
 import { ToolCallContent } from './tool-call-content'
