+++ apps/web/src/features/chat/components/utils/expandable-section.tsx
